builtinhttp.cpp - testCallRefusedAuth(), cast QNetworkReply to make older QCOMPAREs happy

+++ unittests/builtinhttp/builtinhttp.cpp
@@ -169,7 +169,7 @@ private Q_SLOTS:
         client.setAuthentication(auth);
         KDSoapMessage reply = client.call(QLatin1String("getEmployeeCountry"), countryMessage());
         QVERIFY(reply.isFault());
-        QCOMPARE(reply.childValues().child(QLatin1String("faultcode")).value().toInt(), QNetworkReply::AuthenticationRequiredError);
+        QCOMPARE(reply.childValues().child(QLatin1String("faultcode")).value().toInt(), static_cast<int>(QNetworkReply::AuthenticationRequiredError));
     }
 
     void testCorrectHttpHeader()
